Arborescence des besoins presque terminée

+++ ACEM-domain-services/src/main/java/eu/ueb/acem/services/BesoinsReponsesServiceImpl.java
@@ -6,6 +6,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import eu.ueb.acem.dal.DAO;
 import eu.ueb.acem.dal.bleu.BesoinDAO;
@@ -53,27 +54,38 @@ public class BesoinsReponsesServiceImpl implements BesoinsReponsesService {
 	}
 
 	@Override
+	@Transactional
 	public Besoin createOrUpdateBesoin(Long id, String name, Long idParent) {
-		Besoin existingBesoin = besoinDAO.retrieve(id);
-		logger.info("updateBesoin, existingBesoin == {}", existingBesoin);
+		Besoin existingBesoin = besoinDAO.retrieveById(id);
+		logger.info("createOrUpdateBesoin, existingBesoin == {}", existingBesoin);
 		if (existingBesoin == null) {
 			// TODO : ne pas appeler BesoinNode, passer par une factory (?)
 			Besoin newBesoin = besoinDAO.create(new BesoinNode(name));
 			if (idParent != null) {
-				Besoin parent = besoinDAO.retrieve(idParent);
-				newBesoin.addParent(parent);
+				logger.info("createOrUpdateBesoin, idParent == {}", idParent);
+				Besoin parent = besoinDAO.retrieveById(idParent);
+				//newBesoin.addParent(parent);
+				parent.addChild(newBesoin);
+				besoinDAO.update(newBesoin);
+				besoinDAO.update(parent);
 			}
 			return newBesoin;
 		}
 		else {
-			existingBesoin.setNom(name);
+			existingBesoin.setName(name);
 			return besoinDAO.update(existingBesoin);
 		}
 	}
 
 	@Override
+	@Transactional
 	public void deleteBesoin(Long id) {
-		Besoin besoin = besoinDAO.retrieve(id);
-		besoinDAO.delete(besoin);
+		Besoin besoin = besoinDAO.retrieveById(id);
+		if (besoin != null) {
+			besoinDAO.delete(besoin);
+		}
+		else {
+			logger.info("Service deleteBesoin, cannot find besoin with id={}", id);
+		}
 	}
 }
